+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [semantic versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Updated `insert_df_to_hive_table` function in `cdp/io/output.py`; enhanced schema
+  mismatch error logging to show specific column differences.
 
 ### Deprecated